requirements.txt
streamlit
Pillow
simplekml
